- adds muParser original cmakeLists - adds core A.cpp file and coreTest

+++ src/dem_coupling/physicsEngineAdapter/pe/PePhysicsEngineGeometryAdapter.cpp
@@ -7,7 +7,7 @@
 
 PePhysicsEngineGeometryAdapter::PePhysicsEngineGeometryAdapter(walberla::pe::RigidBody* peGeoObject) : peGeoObject(peGeoObject)
 {
-    this->id = peGeoObject->getID();
+    this->id = (int)peGeoObject->getID();
 }
 
 void PePhysicsEngineGeometryAdapter::addForce(const Vector3D& force)
